+++ tests/unit/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from tests.unit.fixtures.adapter import adapter, behavior_flags, config, flags
